requirements.txt

streamlit
numpy
opencv-python-headless
tensorflow
Pillow
